Store metrics in Trainer, save at end.
Previously the reward / Q values were stored as global variables in `test`. This was problematic for multiple reasons (e.g. what if you train + test multiple models?), so now they're stored in the trainer instead. At the end of the training loop, these and the arguments are saved so the results can be plotted later and the model more easily recreated.

Setting the device to CUDA, if available, has been moved out of the trainer and into the args class (in another repo, but it should probably be moved here sometime).